Custom type to handle User.passhash field
Create custom type PasswordType to hash and compare password when doing a
lookup. All the work is done with one way hashes so we don't take any risk
storing or handling plain text passwords.
Because we go with a sha256 hash solution, passhash doesn't need to be bigger
than 64 chararters long.